Refactorización de vistas
- Se agregó la carpeta de utilidades 
- Se refactorizaron algunas vistas en métodos para mayor legibilidad
